feat: Add create token bridge abstraction
